build: Use AX_RUBY_EXT macro
The assumption made for the installation directory RBSIGROK_EXTDIR for
the Ruby bindings only seems to be valid for Debian/Ubuntu based
distributions. A more generic and flexible approach is to use the
AX_RUBY_EXT macro from the autoconf-archive project, which determines
the installation directory from the config variable
`RbConfig::CONFIG['sitearchdir']` and allows to override it at
configure-time using the RUBY_EXT_LIB variable.